requirements.txt
# Django and core dependencies
Django>=3.2,<4.0
django-environ>=0.8.1
django-allauth>=0.47.0
django-cors-headers>=3.10.0
gunicorn>=20.1.0

# Database
psycopg2-binary>=2.9.3

# Azure dependencies
azure-identity>=1.7.0
azure-keyvault-secrets>=4.3.0
azure-storage-blob>=12.9.0
azure-ai-textanalytics>=5.2.0
azure-cognitiveservices-language-textanalytics>=0.2.0
opencensus-ext-azure>=1.1.7
opencensus-ext-django>=0.8.0

# Web scraping and processing
beautifulsoup4>=4.10.0
requests>=2.27.1

# Utils
python-dateutil>=2.8.2
pytz>=2021.3

# Development
python-dotenv>=0.19.2

# Production
whitenoise>=5.3.0  # For serving static files
sentry-sdk>=1.5.1  # For error tracking
